+++ bank/src/main/java/com/operations/bank/request/FundtransferRquest.java
@@ -5,20 +5,20 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class FundtransferRquest {
 	
-	private int fromAccount;
-	private int toAccount;
+	private long fromAccount;
+	private long toAccount;
 	private double amount;
 	private String comments;
-	public int getFromAccount() {
+	public long getFromAccount() {
 		return fromAccount;
 	}
-	public void setFromAccount(int fromAccount) {
+	public void setFromAccount(long fromAccount) {
 		this.fromAccount = fromAccount;
 	}
-	public int getToAccount() {
+	public long getToAccount() {
 		return toAccount;
 	}
-	public void setToAccount(int toAccount) {
+	public void setToAccount(long toAccount) {
 		this.toAccount = toAccount;
 	}
 	public double getAmount() {
